+++ packages/vue/src/components/organisms/SfMegaMenu/SfMegaMenu.stories.js
@@ -4,32 +4,31 @@ import { withKnobs, text } from "@storybook/addon-knobs";
 
 import SfMegaMenu from "./SfMegaMenu.vue";
 
-import SfImage from "../../atoms/SfImage/SfImage.vue";
 import SfHeader from "../SfHeader/SfHeader.vue";
 import SfMenuItem from "../../molecules/SfMenuItem/SfMenuItem.vue";
+import SfBanner from "../../molecules/SfBanner/SfBanner.vue";
+import SfList from "../SfList/SfList.vue";
 
 const AsidePlaceholder = {
-  components: { SfImage },
+  components: { SfBanner },
   data() {
     return {
       isMobile: false,
       tiles: [
         {
-          title: "Last pairs left",
+          title: "THE OFFICE LIFE",
+          subtitle: "T-shirts",
           pictures: {
-            mobile: {
-              url: "assets/storybook/SfMegaMenu/bannerSandals-full.png"
-            },
-            desktop: { url: "assets/storybook/SfMegaMenu/bannerSandals.jpg" }
+            mobile: "assets/storybook/SfMegaMenu/bannerSandals.jpg",
+            desktop: "assets/storybook/SfMegaMenu/bannerSandals.jpg"
           }
         },
         {
-          title: "Beach bags 2=1",
+          title: "ECO SANDALS",
+          subtitle: "T-shirts",
           pictures: {
-            mobile: {
-              url: "assets/storybook/SfMegaMenu/bannerBeachBag-full.png"
-            },
-            desktop: { url: "assets/storybook/SfMegaMenu/bannerBeachBag.jpg" }
+            mobile: "assets/storybook/SfMegaMenu/bannerBeachBag.jpg",
+            desktop: "assets/storybook/SfMegaMenu/bannerBeachBag.jpg"
           }
         }
       ]
@@ -37,9 +36,7 @@ const AsidePlaceholder = {
   },
   computed: {
     root() {
-      return this.isMobile
-        ? {}
-        : { display: "flex", justifyContent: "space-between" };
+      return this.isMobile ? {} : { display: "flex" };
     }
   },
   mounted() {
@@ -55,19 +52,20 @@ const AsidePlaceholder = {
       this.isMobile = event.matches;
     }
   },
-  template: `<div :style="root">
-      <div 
-        v-for="tile in tiles" 
+  template: `
+      <div :style="root">
+      <SfBanner 
+        v-for="(tile, index) in tiles"
         :key="tile.title" 
-        :style="{marginBottom: '1.25rem'}"
-      >
-        <h3 :style="{marginBottom: '1.25rem', textTransform: 'uppercase'}">{{tile.title}}</h3>
-        <SfImage :src="tile.pictures"/>
-      </div>
-    </div>`
+        :title="tile.title" 
+        :subtitle="tile.subtitle" 
+        :image="tile.pictures" 
+        :style="{margin: isMobile ? index === 0 ? '0' : '24px 0 0 0' : index === 0 ? '0' : '0 0 0 24px', '--banner-height': '310px', '--banner-width': '330px'} "
+      />
+      </div>`
 };
 const MegaMenuPlaceholder = {
-  components: { SfMegaMenu, SfMenuItem, AsidePlaceholder },
+  components: { SfMegaMenu, SfMenuItem, AsidePlaceholder, SfList },
   props: {
     title: {
       type: String,
@@ -76,6 +74,9 @@ const MegaMenuPlaceholder = {
     visible: {
       type: Boolean,
       default: false
+    },
+    asideTitle: {
+      default: text("asideTitle", "Featured", "Props")
     }
   },
   data() {
@@ -120,6 +121,7 @@ const MegaMenuPlaceholder = {
   },
   template: `<SfMegaMenu 
         title="Man"
+        :asideTitle="asideTitle"
         :visible="visible"
         :style="{ position: 'absolute', left: 0, width: '100%', top: '100%' }"
       >
@@ -128,11 +130,16 @@ const MegaMenuPlaceholder = {
           :key="subcategory.header" 
           :title="subcategory.header"
         >
-          <SfMenuItem
-            v-for="item in subcategory.items" 
-            :key="item.label" 
-            :label="item.label" 
-          />
+          <SfList>
+            <SfListItem 
+              v-for="item in subcategory.items" 
+              :key="item.label"
+            >
+              <SfMenuItem
+                :label="item.label" 
+              />
+            </SfListItem>
+          </SfList>
         </SfMegaMenuColumn>
         <template #aside>
           <AsidePlaceholder/>
@@ -145,7 +152,8 @@ storiesOf("Organisms|MegaMenu", module)
   .add("Common", () => ({
     components: {
       SfMegaMenu,
-      SfMenuItem
+      SfMenuItem,
+      SfList
     },
     props: {
       title: {
@@ -196,18 +204,19 @@ storiesOf("Organisms|MegaMenu", module)
     template: `<SfMegaMenu 
         :title="title" 
         :visible="visible"
-        :style="{maxWidth: '1240px', margin: 'auto'}"
       >
         <SfMegaMenuColumn 
           v-for="subcategory in subcategories" 
           :key="subcategory.header" 
           :title="subcategory.header"
         >
-          <SfMenuItem
-            v-for="item in subcategory.items" 
-            :key="item.label" 
-            :label="item.label" 
-          />
+          <SfList>
+            <SfListItem v-for="item in subcategory.items" :key="item.label" >
+              <SfMenuItem
+                :label="item.label" 
+              />
+            </SfListItem>
+          </SfList>
         </SfMegaMenuColumn>
       </SfMegaMenu>`
   }))
@@ -215,11 +224,15 @@ storiesOf("Organisms|MegaMenu", module)
     components: {
       SfMegaMenu,
       SfMenuItem,
+      SfList,
       AsidePlaceholder
     },
     props: {
       title: {
         default: text("title", "Man", "Props")
+      },
+      asideTitle: {
+        default: text("asideTitle", "Featured", "Props")
       }
     },
     data() {
@@ -265,19 +278,24 @@ storiesOf("Organisms|MegaMenu", module)
     },
     template: `<SfMegaMenu 
         :title="title"
+        :asideTitle="asideTitle"
         :visible="visible"
-        :style="{maxWidth: '1240px', margin: 'auto'}"
       >
         <SfMegaMenuColumn 
           v-for="subcategory in subcategories" 
           :key="subcategory.header" 
           :title="subcategory.header"
         >
-          <SfMenuItem
-            v-for="item in subcategory.items" 
-            :key="item.label" 
-            :label="item.label"
-          />
+          <SfList>
+            <SfListItem
+              v-for="item in subcategory.items"
+              :key="item.label" 
+            >
+              <SfMenuItem
+                :label="item.label"
+              />
+            </SfListItem>
+          </SfList>
         </SfMegaMenuColumn>
         <template #aside>
           <AsidePlaceholder/>